+++ internal/update/migration_test.go
@@ -3,7 +3,7 @@ package update
 import "testing"
 
 func TestCompatibilityMigration(t *testing.T) {
-	for _, path := range [][2]string{{"1.2.0", "1.3.0"}, {"1.1.0", "1.2.0-rc.1"}, {"1.1.0", "1.2.0-rc.2"}, {"1.2.0-rc.1", "1.2.0-rc.2"}, {"1.2.0-rc.2", "1.2.0-rc.7"}, {"1.2.0-rc.2", "1.2.0"}} {
+	for _, path := range [][2]string{{"1.3.0", "1.3.1"}, {"1.2.0", "1.3.0"}, {"1.1.0", "1.2.0-rc.1"}, {"1.1.0", "1.2.0-rc.2"}, {"1.2.0-rc.1", "1.2.0-rc.2"}, {"1.2.0-rc.2", "1.2.0-rc.7"}, {"1.2.0-rc.2", "1.2.0"}} {
 		m, err := PlanMigration(path[0], path[1])
 		if err != nil || m.Validate() != nil || m.Mutation || !m.PreserveConfiguration || !m.PreserveCredentials || !m.PreserveState {
 			t.Fatalf("migration %v: %+v %v", path, m, err)